ahora si todo lo anterior
