+++ reactApp/package.json
@@ -23,6 +23,7 @@
     "chartjs-plugin-datalabels": "^2.2.0",
     "cors": "^2.8.5",
     "date-fns": "^4.1.0",
+    "dotenv": "^16.4.5",
     "express": "^4.21.1",
     "lucide-react": "^0.447.0",
     "nodemailer": "^6.9.15",
